// src/generated/model/favoriteFilterModel.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FavoriteFilterModel { 
    /**
     * ID of the product
     */
    productId?: string;
    /**
     * IDs of the product
     */
    productsId?: Array<string>;
    /**
     * name of the product
     */
    name?: string;
    /**
     * nutriscore of the product
     */
    nutriscore?: string;
    /**
     * ecoscore of the product
     */
    ecoscore?: string;
    /**
     * ID of the product
     */
    novagroup?: string;
    /**
     * Whether the favorite is active
     */
    isEnabled?: boolean;
    /**
     * Start of creation date range
     */
    dateCreationStart?: string;
    /**
     * End of creation date range
     */
    dateCreationEnd?: string;
}